Se agrega la ruta del Login y se reorganizan los componentes
