+++ frontend/src/components/Section/CardComponent/card-pokemon.css
